Xor fonctionnel pour 1 couche cachée, 2 neurones random de 1 et biais de 1

+++ main.py
@@ -12,7 +12,7 @@ lib = ctypes.cdll.LoadLibrary('./libadd.so')
 def perceptron( hidden_Layer, neurons, random, data, bias, size, lib, nb_Class, prediction, learning_rate):
     func = lib.perceptron
     func.restype = ctypes.c_double
-    func.argtypes = (ctypes.c_int, ctypes.c_int, ctypes.c_int ,ctypes.POINTER(ctypes.c_double), ctypes.c_int, ctypes.c_int, ctypes.c_int, ctypes.POINTER(ctypes.c_int), ctypes.c_double)
+    func.argtypes = (ctypes.c_int, ctypes.c_int, ctypes.c_int ,ctypes.POINTER(ctypes.c_double), ctypes.c_double, ctypes.c_int, ctypes.c_int, ctypes.POINTER(ctypes.c_int), ctypes.c_double)
     data_array = (ctypes.c_double * len(data))(*data)
     prediction_array = (ctypes.c_int * len(prediction))(*prediction)
     res = func(hidden_Layer, neurons, random, data_array, bias ,size, nb_Class, prediction_array, learning_rate)
